Refactoring to fit DeadReckoningService and overall clenaup of MainActivity
