--- DeadCrumbs/app/src/main/java/dead/crumbs/ui/OrientationService.kt
@@ -1,123 +1,0 @@
-package dead.crumbs.ui
-
-import android.app.Service
-import android.content.Context
-import android.content.Intent
-import android.hardware.Sensor
-import android.hardware.SensorEvent
-import android.hardware.SensorEventListener
-import android.hardware.SensorManager
-import android.os.Binder
-import android.os.IBinder
-
-class OrientationService() : Service(), SensorEventListener {
-
-    //Called on creation of OrientationService
-    override fun onCreate() {
-        super.onCreate()
-        setupOrientation()
-    }
-    var callback: ((FloatArray) -> Unit)? = null
-
-    //----------Binding--------------
-    // Binder given to clients
-    private val binder = LocalBinder()
-
-    /**
-     * Class used for the client Binder.  Because we know this service always
-     * runs in the same process as its clients, we don't need to deal with IPC.
-     */
-    inner class LocalBinder : Binder() {
-        // Return this instance of LocalService so clients can call public methods
-        fun getService(): OrientationService = this@OrientationService
-    }
-
-    override fun onBind(intent: Intent): IBinder {
-        return binder
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-
-        // Don't receive any more updates from either sensor.
-        sensorManager.unregisterListener(this)
-    }
-
-    //---------Orientation Sensor Code----------//
-    private lateinit var sensorManager: SensorManager
-    private var sensor: Sensor? = null
-    private val accelerometerReading = FloatArray(3)
-    private val magnetometerReading = FloatArray(3)
-    private val rotationMatrix = FloatArray(9)
-    private val orientationAngles = FloatArray(3)
-
-
-    private fun setupOrientation() {
-        sensorManager = getSystemService(Context.SENSOR_SERVICE) as SensorManager
-        sensor = sensorManager.getDefaultSensor(Sensor.TYPE_GEOMAGNETIC_ROTATION_VECTOR)
-
-        sensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER)?.also { accelerometer ->
-            sensorManager.registerListener(
-                this,
-                accelerometer,
-                SensorManager.SENSOR_DELAY_NORMAL, //TODO change delay?
-                SensorManager.SENSOR_DELAY_UI
-            )
-        }
-        sensorManager.getDefaultSensor(Sensor.TYPE_MAGNETIC_FIELD)?.also { magneticField ->
-            sensorManager.registerListener(
-                this,
-                magneticField,
-                SensorManager.SENSOR_DELAY_NORMAL, //TODO change delay?
-                SensorManager.SENSOR_DELAY_UI
-            )
-        }
-    }
-
-    override fun onAccuracyChanged(sensor: Sensor, accuracy: Int) {
-        // Do something here if sensor accuracy changes.
-        // You must implement this callback in your code.
-    }
-
-    // Compute the three orientation angles based on the most recent readings from
-    // the device's accelerometer and magnetometer.
-    fun updateOrientationAngles() {
-        // Update rotation matrix, which is needed to update orientation angles.
-        SensorManager.getRotationMatrix(
-            rotationMatrix,
-            null,
-            accelerometerReading,
-            magnetometerReading
-        )
-
-        // "mRotationMatrix" now has up-to-date information.
-
-        SensorManager.getOrientation(rotationMatrix, orientationAngles)
-        // "mOrientationAngles" now has up-to-date information, note that the data is in radians.
-
-        callback?.let { it(orientationAngles) }
-    }
-
-    fun updateOrientationAnglesGyroscope(){
-        //Link: https://stackoverflow.com/questions/13350622/getting-rotation-degree-from-gyroscope-android-sensor
-        //You save the rotation matrix from previus measurement, R1.
-        //Then you calculate a new one now, R2.
-        //Then you calculate the following matrix:
-        // R = R1 * transpose(R2)
-        //Then you can calculate the rotation angle as:
-        //angle = acos((trace(R)-1)/2)
-    }
-
-    // Get readings from accelerometer and magnetometer. To simplify calculations,
-    // consider storing these readings as unit vectors.
-    override fun onSensorChanged(event: SensorEvent) {
-        if (event.sensor.type == Sensor.TYPE_ACCELEROMETER) {
-            System.arraycopy(event.values, 0, accelerometerReading, 0, accelerometerReading.size)
-        } else if (event.sensor.type == Sensor.TYPE_MAGNETIC_FIELD) {
-            System.arraycopy(event.values, 0, magnetometerReading, 0, magnetometerReading.size)
-        }
-            updateOrientationAngles()
-    }
-
-
-}
